Make device type register readable

+++ system/src/cpu/ex/load.rs
@@ -54,13 +54,13 @@ impl LoadOperator for Lw {
 
 pub fn lui(_cpu: &mut Cpu, pc: u32, word: u32) -> DcState {
     let rt = ((word >> 16) & 31) as usize;
-    let imm = (word & 0xffff) as i16 as i64;
+    let imm = (word & 0xffff) as i16;
 
     trace!("{:08X}: LUI {}, 0x{:04X}", pc, Cpu::REG_NAMES[rt], imm);
 
     DcState::RegWrite {
         reg: rt,
-        value: imm << 16,
+        value: ((imm as i32) << 16) as i64,
     }
 }
